// src/app/enums/addtask-error.ts
export enum AddTaskTitleError {
    noError = "",
    titleRequired = "This field is required",
    titleLength = "The title exceeds the number of character allowed",
}

export enum AddTaskDescriptionError {
    noError = "",
    descriptionRequired = "This field is required",
    descriptionLength = "The description exceeds the number of characters allowed"
}